+++ common/src/main/java/com/gitlab/srcmc/rctapi/api/ai/utils/MoveType.java
@@ -85,7 +85,15 @@ public enum MoveType {
         Map.<String, Evaluator>entry("meanlook", (from, to, move) -> BattleEffects.Pokemon.State.trapped(to) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("spiderweb", (from, to, move) -> BattleEffects.Pokemon.State.trapped(to) ? 0.0 : 1.0),
         Map.<String, Evaluator>entry("block", (from, to, move) -> BattleEffects.Pokemon.State.trapped(to) ? 0.0 : 1.0),
-        Map.<String, Evaluator>entry("fakeout", (from, to, move) -> BattleStates.get(from.actor.battle).getPokemonState(from).age(BattleEffects.Custom.TURN) > 1 ? 0.0 : 1.0)
+        Map.<String, Evaluator>entry("fakeout", (from, to, move) -> BattleStates.get(from.actor.battle).getPokemonState(from).age(BattleEffects.Custom.TURN) > 1 ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("tailwind", (from, to, move) -> BattleEffects.Side.Tailwind.tailwind(from) ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("gravity", (from, to, move) -> BattleEffects.Field.Gravity.gravity(from) ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("trickroom", (from, to, move) -> BattleEffects.Field.Room.trickroom(from) ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("spikes", (from, to, move) -> (3 - BattleEffects.Side.Hazard.spikes(to))/3.0),
+        Map.<String, Evaluator>entry("stealthrock", (from, to, move) -> BattleEffects.Side.Hazard.stealthrock(to) > 0 ? 0.0 : 1.0),
+        Map.<String, Evaluator>entry("toxicspikes", (from, to, move) -> (2 - BattleEffects.Side.Hazard.toxicspikes(to))/2.0),
+        Map.<String, Evaluator>entry("stickyweb", (from, to, move) -> BattleEffects.Side.Hazard.stickyweb(to) > 0 ? 0.0 : 1.0)
+        // Map.<String, Evaluator>entry("spikes", (from, to, move) -> BattleEffects.Side.Hazard.sharpsteel(to) > 0 ? 0.0 : 1.0),
     );
 
     // custom move event handlers